Improved reel icon.  Made skip annotations on edl.

+++ mrViewer/src/gui/mrvImageBrowser.cpp
@@ -3867,7 +3867,7 @@ void ImageBrowser::next_image_limited()
     int v = value();
 
     ++v;
-    if ( size_t(v) >= (int)reel->images.size() )
+    if ( size_t(v) >= reel->images.size() )
     {
         if ( play ) view()->play(play);
         return;
